fix(ui): improve markdown editor indentation counting
Previously, the markdown editor would count instances of indentation
text (ie. 4 spaces, tabs, and "> ") from anywhere in the line, rather
than only counting those at the start of the line, so attempting to
indent a line of text containing those strings in other locations could
fail when it ought to be indented. This commit replaces the previous
regex implementation with a function to count indents at the start of
the line instead.