chore: add `GetShortStat`
Add a new function to the suite of getting shortstat, with the most
straight forward name yet, `GetShortStat`. The main difference being
that it doesn't retry on "no merge base found" and that it uses
`git-diff-tree` instead of `git-diff` which has better performance on
certain types of diff. It also detects _common_ errors and returns a
dedicated error for them so callers of this function don't have to try
and parse the error themselves.