feat: simplify `GetPullRequestFiles`
`GetPullRequestFiles` is the API route handler to get the files that
are changed in a pull request, it has to know the startcommit and end
commit to diff for (so it can gather the changed file and other
information). The end commit is clear, the pr ref (`ref/pull/xxx/head`).
However the start commit has to be computed, it is the merge base commit
between the base branch and pr ref. However if the pr was merged, then
we should use the `pr.MergeBase` as it's possible the `pr.BaseBranch` no
longer exists.

Instead of doing this computation via `GetCompareInfo` that also does
some other computations, compute the merge base directly ourselves in
this function, if no merge base exists then fallback to the base
reference (this is the same behavior as in `GetCompareInfo`). The only
differene is that in the case of the fallbakc we don't convert the base
ref to a commit ID, this is not necessary as the call to `git-diff` will
accept any valid reference.