chore(ui): remove pageData fallback value
This fallback value was added in Gitea a long time ago, presumably
before there was a guarantee that context.go would always set certain
data in pageData. Now, there are a number of places in the code which
access keys of pageData safely without checking for their existence, so
this fallback ought not be needed anymore.