feat: make create tag/branch of commit native dialog
There's a simplification going on here, the information inside the modal
is constructed by some javascript code upon opening the modal. There's
no need for that as we know the values already so lots of attributes are
removed and the dialog contains all the information now.

`show-modal` is still using the fomantic modals but works correctly with
`<dialog>` elements.

The modal is moved to the native `<dialog>` element.