Enforce 1-based rule for pages that are <= 1
At `/routers/api/v1/utils/page.go`

The function `GetListOptions` was extracting page with
ctx.FormInt("page"), essentially breaking the rule of 1-based in case
the query param page was not provided, since golang will initialize int
to 0.

With this change, a method `ToCorrectPage` to enforce the rule was added
and Page is extracted using this method instead.