fix css selectors
unfortunately our avatars css is a mess:
`.avatar` is used for both `a` container and for `img/svg` inside of it.

however, the `img/svg` always have `.ui` alongside `.avatar`. that's how i determined what is for what and patched selectors in problematic areas i could find.

i tried different approaches and settled on this one as least messy and districutive.