chore(ui): make danger buttons brighter
More specifically, the new lightness 0.26 is same as --color-box-body instead of being darker than it, so it doesn't break the rule that elements inside of a form/box shouldn't be darker than the background they're on, which I made up.

FF contrast calc: 5.01 (AA) which I don't understand how it got because for text w/ same lightness on background with same lightness but with steel hue it only scores 4.88 (also AA).

Anyway, the new color looks good in person, still enough contrast, and same brightness should prevent it from looking wrong.