fix: show spinner when loading content history menu
- When clicking on the content history menu "edited" item, show a
spinner animation as before the dropdown is show can take a measurable
amount of time.
- We cannot rely on fomantic adding the loading indicator (although it
does have this capability), it only adds this indicator when waiting for
the network request. For a fast Forgejo instance waiting for the network
response is faster than what fomantic is doing to prepare the dropdown.
- Resolves forgejo/forgejo#9841