fix: discoverability of team invites by email
When adding someone to a team, it is possible to select
them either by their username or by email address, in which
case they will get an invite by email. There is no hint of
that feature in the UI, so I propose to improve the placeholder
of this field to make this more discoverable.

Related to #8951.