fix: update devcontainer tag to go:1.25-trixie
In a recent commit (188810c9e9), the devcontainer base image was moved
to 2.0-bullseye.

This tag only includes Go 1.24. With the recent move to Go 1.25, this
results in the following build failure when downloading deps:

```
/usr/local/go/bin/go mod download
go: go.mod requires go >= 1.25.0 (running go 1.24.6; GOTOOLCHAIN=local)
```

Let's set this to `1.25-trixie` to match our current Go version and fix
this error.