feat: add migration to keying
The most intense part, the two previous migrations to keying resulted in
regressions. The bugs have been ironed out and we know our expectations
regarding the state of database utilities in the codebase.

The migration closely follows the previous migration, it uses a
transaction and uses db.Iterate - if we cannot decrypt then we can
safely delete the row.