chore: simplify `GetNote`
Return the Note object (avoid C-style functions).

Motivation to refactor this function is to avoid the function that uses
last commit cache for git-notes, because it is not needed at the scale
of git-notes. In the worst case it can be considered to make a patch to
git to get the message and commitID, because git seems to have efficient
code to do this (for getting messages, but does not expose the commit id).