Review feedback: add tags, one step delete, comment delete
Added the indexes as tags on the model fields and switched
the delete to be done in one step instead of iteratively.

Also restored the method that updates is_deleted in the action
table to true in the comment model and switched it
to directly delete the related action instead.