Restore indices function
Restored the indices function because it is likely
the new indexes may differ from the the old ones,
potentially leading to the fields being in a different
order in the indexes.

Since the order of fields in an index can have a significant
performance impact, we are restoring the function.