fix: garbage collect lingering actions logs
If, for any reason (e.g. server crash), a task is recorded as done
in the database but the logs are still in the database instead of
being in storage, they need to be collected.

The log_in_storage field is only set to true after the logs have been
transfered to storage and can be relied upon to reflect which tasks
have lingering logs.

A task collects lingering logs every day, 3000 at a time, sleeping
one second between them. In normal circumstances there will be only a
few of them, even on a large instance, and there is no need to collect
them as quickly as possible.

When there are a lot of them for some reason, garbage collection must
happen at a rate that is not too hard on storage I/O.

Refs https://codeberg.org/forgejo/forgejo/issues/9999

(cherry picked from commit 9982b2eea46069fa7a29b0f186dd0fac6494f173)