chore: remove webkit and mobile safari from playwright
They are comically unreliable, will fail for unexplainable reasons and
are very hard to run locally in comparisions with the other supported
platforms. I do not remember the last time where these two platforms
were able to catch a regression where the other platforms did not.

I would like to stress, for the historical record, that many hours has
been devoted into adjusting the tests and following best practices to
make these two platforms but despite those, IMO wasted, efforts these
two platforms are causing many hours of wasted CPU time simply because
they are flaky and make (new) contributors nervous if their change
contains a regression or not.

To my knowledge, the tests are not broken for these two platforms. If
you go to the issue tracker you will not find many issues by users that
use these two platforms and report that Forgejo is broken. It does not
reflect reality.

This is the sunk cost fallacy, bite the bullet and agree that these
platforms will not contribute postively to Forgejo's excellent test
suite.