user code verification
implement the User Interaction part of RFC 8628 (Section 3.3).
Comprised of two endpoints that require a logged-in session:

Initially, the `device_authorization` response directs the user to
visit `/login/oauth/device`, where the user is prompted to enter   the
user code. On the second page, the user is shown details about   the
application they're about to authorize, and asked to confirm the
authorization.

If the client application implements the "Non-Textual Verification URI
Optimization" (Sectino 3.3.1), the URL the user visists will include
a `user_code` query parameter. If this is present, we skip the initial
page asking for the code to be entered.

By confirming or denying the authorization requests, the user visits
`/login/oauth/grant_device`. Depending on the user's decision, the
endpoint either upserts a Grant for the user/app combination, or clears
out the user_code, thus invalidating the device authorization.

See: https://www.rfc-editor.org/rfc/rfc8628#section-3.3
See: https://www.rfc-editor.org/rfc/rfc8628#section-3.3.1