fix(ui): fix issue with readme being wider than container
The readme on a repos home page was 2px wider than the other elements in the main container.

This change also has flow on effects for other ui elements using the same css classes, which I have tried to address, but it's hard to be certain that I have caught all of them.