chore(lint): Enforce consistent import aliases for services and models
To make sure that the code stays maintainable, I added the importas
linter to ensure that the imports for models and services stay
consistent.

I realised that this might be needed after finding some discrepancies
between singular/plural naming, and, especially in the case of the
`forgejo.org/services/context` package, multiple different aliases like
`gitea_ctx`, `app_context` and `forgejo_context`. I decided for
`app_context`, as that seems to be the most commonly used naming.

Signed-off-by: nachtjasmin <nachtjasmin@posteo.de>