fix(ui): URL-encode login provider name in the `href` attribute
The authentication provider's name ($provider.DisplayName) is not URL-encoded, so any illegal characters (e.g., '/') will be put in the href attribute verbatim.
For example, if the provider's name is `foo/bar`, the href attribute will link to `/user/oauth2/foo/bar` instead of `/user/oauth2/foo%2Fbar`, resulting in a 404 error.

This commit fixes this behaviour by URL-encoding the provider's DisplayName before appending it to the href attribute.

Signed-off-by: doasu <me@doasu.dev>