fix: include remote users when counting users
It matters when checking if an admin user is the last one or not.

The driver compliance tests create an admin user that is of type
remote. When trying to delete it at cleanup, it will fail because the
count is 1 instead of 2 and the deletion fails because you are not
allowed to delete the last admin user.