feat: allow placeholders in group team mappings
Currently, Forgejo supports configuring static group team mappings for
an authentication source (e.g., OIDC) that map groups received from the
authentication source to Forgejo organizations and teams. For example,
the following mapping

```json
{"Developer": {"MyForgejoOrganization": ["MyForgejoTeam1", "MyForgejoTeam2"]}}
```

automatically adds a user in the OIDC group `Developer` to the teams
`MyForgejoTeam1` and `MyForgejoTeam2` in organization
`MyForgejoOrganization`.

In order to make such mappings more dynamic and to avoid having to
update the mappings for each new group, add support for placeholders in
the mappings like in the following example:

```json
{
    "group-{org}-{team}": {"{org}": ["{team}"]},
    "other-{org}-{team}": {"other-{org}": ["other-{team}"]},
    "fixed-{org}": {"{org}-fixed": ["team3", "team4"]}
}
```

In this example, the mapping adds a user in groups `group-org1-team1`,
`other-org2-team2`, and `fixed-org3` to team `team1` in organization
`org1`, to team `other-team2` in organization `other-org2`, and to teams
`team3` and `team4` in organization `org3-fixed`.

Additionally, this changes the current behaviour when removing users
from teams: teams to remove cannot be read directly from mappings with
placeholders. Instead, get all organizations and teams of a user and
remove teams that are not added via a mapping. Thus, the user is only in
teams that are added via such a mapping and no other teams.

Further notes:

- Only tested with OIDC
- Adding users to teams becomes more complex with regexes
- Removing users from teams becomes more complex with DB accesses
- Admins must be more careful when specifying mappings with placeholders to
  avoid conflicts or unwanted effects. For example, organization and team names
  should not contain the separator `-` in the example above.

Signed-off-by: hwipl <hwipl@noreply.codeberg.org>