fix: create branch create event before event for pushed commits
Currently, when a user creates a new branch locally and pushes it
afterwards to the remote repository, Forgejo first creates events for
the commits and then creates a branch event.
This means that there is an event saying: "X pushed to branch Y" and
then an event saying: "X created branch Y" which is confusing.